Improve DocumentBasedFixAllProvider performance for multiple documents
Previously, custom fix all providers called FixAllContext.GetDocumentDiagnosticsAsync
for every document in the fix all scope. This change replaces those calls
with a single call to FixAllContextHelpers.GetDocumentDiagnosticsToFixAsync,
which was extracted from CustomBatchFixAllProvider. The result is improved
performance when fixing violations in large projects, especially when only
a few documents in the project contained violations.
