+++ StyleCop.Analyzers/StyleCop.Analyzers.CodeFixes/Helpers/FixAllContextHelper.cs
@@ -1,0 +1,156 @@
+﻿// Copyright (c) Tunnel Vision Laboratories, LLC. All Rights Reserved.
+// Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
+
+namespace StyleCop.Analyzers.Helpers
+{
+    using System.Collections.Concurrent;
+    using System.Collections.Immutable;
+    using System.Linq;
+    using System.Threading;
+    using System.Threading.Tasks;
+    using Microsoft.CodeAnalysis;
+    using Microsoft.CodeAnalysis.CodeFixes;
+
+    internal static class FixAllContextHelper
+    {
+        public static async Task<ImmutableDictionary<Document, ImmutableArray<Diagnostic>>> GetDocumentDiagnosticsToFixAsync(FixAllContext fixAllContext)
+        {
+            var allDiagnostics = ImmutableArray<Diagnostic>.Empty;
+            var projectsToFix = ImmutableArray<Project>.Empty;
+
+            var document = fixAllContext.Document;
+            var project = fixAllContext.Project;
+
+            switch (fixAllContext.Scope)
+            {
+            case FixAllScope.Document:
+                if (document != null)
+                {
+                    var documentDiagnostics = await fixAllContext.GetDocumentDiagnosticsAsync(document).ConfigureAwait(false);
+                    return ImmutableDictionary<Document, ImmutableArray<Diagnostic>>.Empty.SetItem(document, documentDiagnostics);
+                }
+
+                break;
+
+            case FixAllScope.Project:
+                projectsToFix = ImmutableArray.Create(project);
+                allDiagnostics = await fixAllContext.GetAllDiagnosticsAsync(project).ConfigureAwait(false);
+                break;
+
+            case FixAllScope.Solution:
+                projectsToFix = project.Solution.Projects
+                    .Where(p => p.Language == project.Language)
+                    .ToImmutableArray();
+
+                var diagnostics = new ConcurrentDictionary<ProjectId, ImmutableArray<Diagnostic>>();
+                var tasks = new Task[projectsToFix.Length];
+                for (int i = 0; i < projectsToFix.Length; i++)
+                {
+                    fixAllContext.CancellationToken.ThrowIfCancellationRequested();
+                    var projectToFix = projectsToFix[i];
+                    tasks[i] = Task.Run(
+                        async () =>
+                        {
+                            var projectDiagnostics = await fixAllContext.GetAllDiagnosticsAsync(projectToFix).ConfigureAwait(false);
+                            diagnostics.TryAdd(projectToFix.Id, projectDiagnostics);
+                        }, fixAllContext.CancellationToken);
+                }
+
+                await Task.WhenAll(tasks).ConfigureAwait(false);
+                allDiagnostics = allDiagnostics.AddRange(diagnostics.SelectMany(i => i.Value));
+                break;
+            }
+
+            if (allDiagnostics.IsEmpty)
+            {
+                return ImmutableDictionary<Document, ImmutableArray<Diagnostic>>.Empty;
+            }
+
+            return await GetDocumentDiagnosticsToFixAsync(allDiagnostics, projectsToFix, fixAllContext.CancellationToken).ConfigureAwait(false);
+        }
+
+        public static async Task<ImmutableDictionary<Project, ImmutableArray<Diagnostic>>> GetProjectDiagnosticsToFixAsync(FixAllContext fixAllContext)
+        {
+            var project = fixAllContext.Project;
+            if (project != null)
+            {
+                switch (fixAllContext.Scope)
+                {
+                case FixAllScope.Project:
+                    var diagnostics = await fixAllContext.GetProjectDiagnosticsAsync(project).ConfigureAwait(false);
+                    return ImmutableDictionary<Project, ImmutableArray<Diagnostic>>.Empty.SetItem(project, diagnostics);
+
+                case FixAllScope.Solution:
+                    var projectsAndDiagnostics = new ConcurrentDictionary<Project, ImmutableArray<Diagnostic>>();
+                    var options = new ParallelOptions() { CancellationToken = fixAllContext.CancellationToken };
+                    Parallel.ForEach(project.Solution.Projects, options, proj =>
+                    {
+                        fixAllContext.CancellationToken.ThrowIfCancellationRequested();
+                        var projectDiagnosticsTask = fixAllContext.GetProjectDiagnosticsAsync(proj);
+                        projectDiagnosticsTask.Wait(fixAllContext.CancellationToken);
+                        var projectDiagnostics = projectDiagnosticsTask.Result;
+                        if (projectDiagnostics.Any())
+                        {
+                            projectsAndDiagnostics.TryAdd(proj, projectDiagnostics);
+                        }
+                    });
+
+                    return projectsAndDiagnostics.ToImmutableDictionary();
+                }
+            }
+
+            return ImmutableDictionary<Project, ImmutableArray<Diagnostic>>.Empty;
+        }
+
+        private static async Task<ImmutableDictionary<Document, ImmutableArray<Diagnostic>>> GetDocumentDiagnosticsToFixAsync(
+            ImmutableArray<Diagnostic> diagnostics,
+            ImmutableArray<Project> projects,
+            CancellationToken cancellationToken)
+        {
+            var treeToDocumentMap = await GetTreeToDocumentMapAsync(projects, cancellationToken).ConfigureAwait(false);
+
+            var builder = ImmutableDictionary.CreateBuilder<Document, ImmutableArray<Diagnostic>>();
+            foreach (var documentAndDiagnostics in diagnostics.GroupBy(d => GetReportedDocument(d, treeToDocumentMap)))
+            {
+                cancellationToken.ThrowIfCancellationRequested();
+                var document = documentAndDiagnostics.Key;
+                var diagnosticsForDocument = documentAndDiagnostics.ToImmutableArray();
+                builder.Add(document, diagnosticsForDocument);
+            }
+
+            return builder.ToImmutable();
+        }
+
+        private static async Task<ImmutableDictionary<SyntaxTree, Document>> GetTreeToDocumentMapAsync(ImmutableArray<Project> projects, CancellationToken cancellationToken)
+        {
+            var builder = ImmutableDictionary.CreateBuilder<SyntaxTree, Document>();
+            foreach (var project in projects)
+            {
+                cancellationToken.ThrowIfCancellationRequested();
+                foreach (var document in project.Documents)
+                {
+                    cancellationToken.ThrowIfCancellationRequested();
+                    var tree = await document.GetSyntaxTreeAsync(cancellationToken).ConfigureAwait(false);
+                    builder.Add(tree, document);
+                }
+            }
+
+            return builder.ToImmutable();
+        }
+
+        private static Document GetReportedDocument(Diagnostic diagnostic, ImmutableDictionary<SyntaxTree, Document> treeToDocumentsMap)
+        {
+            var tree = diagnostic.Location.SourceTree;
+            if (tree != null)
+            {
+                Document document;
+                if (treeToDocumentsMap.TryGetValue(tree, out document))
+                {
+                    return document;
+                }
+            }
+
+            return null;
+        }
+    }
+}
